Move failOnError to helpers package and import helper package instead
