+++ capabilityreceiver/cmd/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 
-	"github.com/onc-healthit/lantern-back-end/endpointmanager/sharedfunctions"
+	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/helpers"
 
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/config"
 
@@ -17,21 +17,21 @@ import (
 
 func main() {
 	err := config.SetupConfig()
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 
 	store, err := postgresql.NewStore(viper.GetString("dbhost"), viper.GetInt("dbport"), viper.GetString("dbuser"), viper.GetString("dbpassword"), viper.GetString("dbname"), viper.GetString("dbsslmode"))
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 	log.Info("Successfully connected to DB!")
 
 	// Set up the queue for sending messages
 	qName := viper.GetString("capquery_qname")
 	messageQueue, channelID, err := accessqueue.ConnectToServerAndQueue(viper.GetString("quser"), viper.GetString("qpassword"), viper.GetString("qhost"), viper.GetString("qport"), qName)
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 	log.Info("Successfully connected to Queue!")
 	defer messageQueue.Close()
 
 	ctx := context.Background()
 
 	err = capabilityhandler.ReceiveCapabilityStatements(ctx, store, messageQueue, channelID, qName)
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 }
