+++ endpointmanager/cmd/chplquerier/main.go
@@ -10,7 +10,7 @@ import (
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/chplquerier"
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/config"
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/endpointmanager/postgresql"
-	"github.com/onc-healthit/lantern-back-end/endpointmanager/sharedfunctions"
+	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/helpers"
 	log "github.com/sirupsen/logrus"
 
 	"github.com/spf13/viper"
@@ -20,10 +20,10 @@ func main() {
 	var err error
 
 	err = config.SetupConfig()
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 
 	store, err := postgresql.NewStore(viper.GetString("dbhost"), viper.GetInt("dbport"), viper.GetString("dbuser"), viper.GetString("dbpassword"), viper.GetString("dbname"), viper.GetString("dbsslmode"))
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 	defer store.Close()
 	log.Info("Successfully connected!")
 
@@ -43,9 +43,9 @@ func main() {
 	log.Infof("user agent is %s", userAgent)
 
 	err = chplquerier.GetCHPLCriteria(ctx, store, client, userAgent)
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 	err = chplquerier.GetCHPLVendors(ctx, store, client, userAgent)
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 	err = chplquerier.GetCHPLProducts(ctx, store, client, userAgent)
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 }
