+++ endpointmanager/cmd/endpointlinker/main.go
@@ -7,7 +7,7 @@ import (
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/config"
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/endpointlinker"
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/endpointmanager/postgresql"
-	"github.com/onc-healthit/lantern-back-end/endpointmanager/sharedfunctions"
+	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/helpers"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
 )
@@ -20,13 +20,13 @@ func main() {
 	log.Info("Starting to link FHIR endpoints to npi organizations")
 
 	err := config.SetupConfig()
-	sharedfunctions.failOnError("Error setting up config", err)
+	helpers.FailOnError("Error setting up config", err)
 	ctx := context.Background()
 
 	store, err := postgresql.NewStore(viper.GetString("dbhost"), viper.GetInt("dbport"), viper.GetString("dbuser"), viper.GetString("dbpassword"), viper.GetString("dbname"), viper.GetString("dbsslmode"))
-	sharedfunctions.failOnError("Error creating store", err)
+	helpers.FailOnError("Error creating store", err)
 
 	err = endpointlinker.LinkAllOrgsAndEndpoints(ctx, store, verbose)
-	sharedfunctions.failOnError("Error linking all orgs and enpoints", err)
+	helpers.FailOnError("Error linking all orgs and enpoints", err)
 
 }
