+++ endpointmanager/cmd/nppescontactpopulator/main.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"os"
 
-	"github.com/onc-healthit/lantern-back-end/endpointmanager/sharedfunctions"
+	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/helpers"
 
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/config"
 
@@ -17,20 +17,20 @@ import (
 
 func main() {
 	err := config.SetupConfig()
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 
 	ctx := context.Background()
 	store, err := postgresql.NewStore(viper.GetString("dbhost"), viper.GetInt("dbport"), viper.GetString("dbuser"), viper.GetString("dbpassword"), viper.GetString("dbname"), viper.GetString("dbsslmode"))
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 	if len(os.Args) != 2 {
 		log.Fatal("NPPES contact csv file not provided as argument.")
 	}
 	fname := os.Args[1]
 	err = store.DeleteAllNPIContacts(ctx)
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 
 	log.Info("Adding NPI FHIR URLs to database...")
 	added, err := nppesquerier.ParseAndStoreNPIContactsFile(ctx, fname, store)
 	log.Infof("Added %d NPI FHIR URLs to database\n", added)
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 }
