+++ endpointmanager/pkg/helpers/helpers.go
@@ -1,6 +1,9 @@
 package helpers
 
-import "sort"
+import(
+	"sort"
+	"log"
+)
 
 // StringArrayContains checks if the string array contains the provided string.
 func StringArrayContains(l []string, s string) bool {
@@ -41,4 +44,15 @@ func StringArraysEqual(l1 []string, l2 []string) bool {
 		}
 	}
 	return true
+}
+
+// Fail on error checks if err is not equal to nil and if it isn't, logs failure and exits the program
+func FailOnError(errString string, err error) {
+	if err != nil {
+		if errString == "" {
+			log.Fatalf("%s", err)
+		} else {
+			log.Fatalf("%s %s", errString, err)
+		}
+	}
 }
