+++ lanternmq/test/sendTopic/send.go
@@ -9,7 +9,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/onc-healthit/lantern-back-end/endpointmanager/sharedfunctions"
+	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/helpers"
 	"github.com/onc-healthit/lantern-back-end/lanternmq"
 	"github.com/onc-healthit/lantern-back-end/lanternmq/rabbitmq"
 )
@@ -41,16 +41,16 @@ func main() {
 	defer mq.Close()
 
 	err := mq.Connect("guest", "guest", "localhost", "5672")
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 	ch, err := mq.CreateChannel()
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 
 	err = mq.DeclareExchange(ch, "logs_topic", "topic")
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 
 	body := bodyFrom(os.Args)
 	severity := severityFrom(os.Args)
 	err = mq.PublishToExchange(ch, "logs_topic", severity, body)
-	sharedfunctions.FailOnError("", err)
+	helpers.FailOnError("", err)
 	log.Printf(" [x] Sent %s", body)
 }
